Add a microservice for product catalog and checkout
This commit adds two micro-services. The first service is a "product catalog" service that can retrieve a list of products and their names. The second service is a "checkout" service which can retrieve a list of "orders" and the products that were inside of these orders. This demonstrates the concept of federation, as the order that is queried inside the "checkout" service simply returns a reference representation of the "product" (with the ID only). The "catalog" service is then responsible for resolving the references to the products by their IDs.