Add a separate database to each of the microservices rather than a static array of products and orders
This commit adds a database for each of the services, as well as a ORM (ie: Object Relational Mapper) to make queries to the databases. Right now, these databases are automatically created on application start-up, but in a production-like environment this would not be a good idea. Still, it works fine for testing.