Add authentication
This commit adds authentication via an accounts service, which prevents users from querying orders that they did not personally make. We validate incoming tokens via the gateway and forward them to services via the `x-kinabalu-jwt-payload` header. It is important to remember that his channel is secure, since our services our not reachable directly over the internet.