Add the ability to place an order
Users can now place orders for a list of products. This commit demonstrates how to directly query a service from another service (ie: checkout service querying the catalog service for pricing information during the checkout flow). This can be necessary when you require the data within a mutation or similar, rather than the capability to return a reference to the data.